+++ multibody/contact_solvers/BUILD.bazel
@@ -179,8 +179,8 @@ drake_cc_library(
 
 drake_cc_library(
     name = "supernodal_solver",
-    srcs = ["supernodal_solver.cc"],
-    hdrs = ["supernodal_solver.h"],
+    srcs = ["conex_supernodal_solver.cc"],
+    hdrs = ["conex_supernodal_solver.h"],
     interface_deps = [
         ":matrix_block",
         "//common:essential",
