+++ multibody/contact_solvers/sap/sap_solver.h
@@ -5,9 +5,9 @@
 #include <utility>
 #include <vector>
 
+#include "drake/multibody/contact_solvers/conex_supernodal_solver.h"
 #include "drake/multibody/contact_solvers/sap/sap_model.h"
 #include "drake/multibody/contact_solvers/sap/sap_solver_results.h"
-#include "drake/multibody/contact_solvers/supernodal_solver.h"
 #include "drake/systems/framework/context.h"
 
 namespace drake {
